(issue #384) Alter AbstractBridgeService to ensure LoggingFilter is always placed after the codec, even when trace (message) level log is not enabled, so codec exceptions will always be logged by the logging filter.
