+++ transport/spi/src/main/java/org/kaazing/gateway/transport/AbstractBridgeService.java
@@ -29,17 +29,11 @@ import org.apache.mina.util.ExceptionMonitor;
 import org.slf4j.Logger;
 import org.kaazing.gateway.transport.LoggingFilter;
 
-import org.jboss.netty.channel.socket.nio.NioWorker;
-import org.kaazing.gateway.transport.ExceptionLoggingFilter;
-import org.kaazing.gateway.transport.ObjectLoggingFilter;
 import org.kaazing.mina.core.service.AbstractIoServiceEx;
 import org.kaazing.mina.core.service.IoProcessorEx;
 import org.kaazing.mina.core.session.IoSessionConfigEx;
-import org.kaazing.mina.netty.util.threadlocal.VicariousThreadLocal;
 
 public abstract class AbstractBridgeService<T extends AbstractBridgeSession<?, ?>> extends AbstractIoServiceEx implements BridgeService {
-    private static final String LOGGING_FILTER_NAME_SUFFIX = "#logging";
-   
 
     private IoProcessorEx<T> processor;
 
@@ -109,6 +103,9 @@ public abstract class AbstractBridgeService<T extends AbstractBridgeSession<?, ?
         if (logger != null) {
             // setup logging filters for the new session
             LoggingFilter.addIfNeeded(logger, session, getTransportMetadata().getName());
+            if (session.getParent() != null) {
+                LoggingFilter.moveAfterCodec(session.getParent());
+            }
         }
     }
